requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
openai==2.32.0
jinja2==3.1.6
python-multipart==0.0.26
pydantic==2.13.3
python-dotenv==1.2.2
pytest==9.0.3
pytest-timeout==2.4.0
aiofiles==25.1.0
